refactor: notification_senderパッケージ内のProviderをGenerator Providerに置き換え
