+++ packages/notification_sender/lib/notification_sender.dart
@@ -4,5 +4,4 @@ library notification_sender;
 export 'package:flutter_local_notifications/flutter_local_notifications.dart'
     show PendingNotificationRequest;
 
-export 'src/notification_sender.dart';
 export 'src/notification_sender_provider.dart';
