+++ packages/notification_sender/test/src/notification_sender_provider_test.mocks.dart
@@ -1,5 +1,5 @@
 // Mocks generated by Mockito 5.4.0 from annotations
-// in notification_sender/test/src/notification_sender_test.dart.
+// in notification_sender/test/src/notification_sender_provider_test.dart.
 // Do not manually edit this file.
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
